simplifying prepare_measurement
