+++ eds_scikit/biology/utils/prepare_measurement.py
@@ -15,9 +15,10 @@ from eds_scikit.biology.utils.process_measurement import (
     filter_measurement_valid,
     normalize_unit,
     tag_measurement_anomaly,
+    convert_measurement_units
 )
 from eds_scikit.io.settings import mapping
-from eds_scikit.utils.framework import is_koalas, to
+from eds_scikit.utils.framework import is_koalas
 from eds_scikit.utils.typing import Data, DataFrame
 
 
@@ -80,80 +81,10 @@ def prepare_measurement_table(
         logger.info(
             "Lazy preparation not available if convert_units=True. Computed table will be cached."
         )
-        if is_koalas(measurement):
-            measurement.cache()
-            conversion_table = to(
-                "koalas", get_conversion_table(measurement, concept_sets)
-            )
-        else:
-            conversion_table = get_conversion_table(measurement, concept_sets)
+        measurement = convert_measurement_units(measurement, concept_sets)
 
-        measurement = measurement.merge(
-            conversion_table, on=["concept_set", "unit_source_value"]
-        )
-        measurement["value_as_number_normalized"] = (
-            measurement["value_as_number"] * measurement["factor"]
-        )
     if is_koalas(measurement):
         measurement.cache()
         logger.info("Done. Once computed, measurement will be cached.")
 
     return measurement
-
-
-def get_conversion_table(
-    measurement: DataFrame, concepts_sets: List[ConceptsSet]
-) -> DataFrame:
-
-    """Given measurement dataframe and list of concepts_sets output conversion table to be merged with measurement.
-
-    Parameters
-    ----------
-    measurement : DataFrame
-    concepts_sets : List[ConceptsSet]
-
-    Returns
-    -------
-    DataFrame
-        Conversion table to be merged with measurement
-    """
-    conversion_table = (
-        measurement.groupby("concept_set")["unit_source_value"]
-        .unique()
-        .explode()
-        .to_frame()
-        .reset_index()
-    )
-    conversion_table = to("pandas", conversion_table)
-    conversion_table["unit_source_value_normalized"] = conversion_table[
-        "unit_source_value"
-    ]
-    conversion_table["factor"] = conversion_table.apply(
-        lambda x: 1 if x.unit_source_value_normalized else 0, axis=1
-    )
-
-    for concept_set in concepts_sets:
-        unit_source_value_normalized = concept_set.units.target_unit
-        conversion_table.loc[
-            conversion_table.concept_set == concept_set.name,
-            "unit_source_value_normalized",
-        ] = conversion_table.apply(
-            lambda x: unit_source_value_normalized
-            if concept_set.units.can_be_converted(
-                x.unit_source_value, unit_source_value_normalized
-            )
-            else concept_set.units.get_unit_base(x.unit_source_value),
-            axis=1,
-        )
-        conversion_table.loc[
-            conversion_table.concept_set == concept_set.name, "factor"
-        ] = conversion_table.apply(
-            lambda x: concept_set.units.convert_unit(
-                x.unit_source_value, x.unit_source_value_normalized
-            ),
-            axis=1,
-        )
-
-    conversion_table = conversion_table.fillna(1)
-
-    return conversion_table
